:bug: Correctly encode image paths for src attribute
Images such as '?.png', '#.png' and '%2F.png' all fail loading due to not being
correctly encoded before passing off to the img src attribute. The fix for this
is to encode the URI using encodeURI and then manually encode '?' and '#' since
file paths cannot have a hash fragments or query string.

Fixes #38